Rename parser wrapper
Because upstream did that too
