+++ query/search_original.js
@@ -1,7 +1,7 @@
 const _ = require('lodash');
 const peliasQuery = require('pelias-query');
 var defaults = require('./search_defaults');
-const textParser = require('./text_parser_addressit');
+const textParser = require('./text_parser_pelias');
 const config = require('pelias-config').generate().api;
 
 var placeTypes = require('../helper/placeTypes');
